Add files via upload
The region was hard coded.  Have made changes so that the standard environment variable AWS_DEFAULT_REGION is used. If this variable is not set, then the AWS default region for the profile is used. If the default region is not set, the user is invited to key in the region.